+++ src/Dumper/Sql/Metadata/MysqlMetadata.php
@@ -49,26 +49,46 @@ class MysqlMetadata implements MetadataInterface
      */
     public function getForeignKeys(string $tableName): array
     {
-        $query = 'SELECT CONSTRAINT_NAME, TABLE_NAME, GROUP_CONCAT(COLUMN_NAME) AS COLUMN_NAMES, '
-            . 'REFERENCED_TABLE_NAME, GROUP_CONCAT(REFERENCED_COLUMN_NAME) AS REFERENCED_COLUMN_NAMES '
+        $query = 'SELECT CONSTRAINT_NAME, TABLE_NAME, COLUMN_NAME, '
+            . 'REFERENCED_TABLE_NAME, REFERENCED_COLUMN_NAME '
             . 'FROM INFORMATION_SCHEMA.KEY_COLUMN_USAGE '
             . 'WHERE REFERENCED_TABLE_NAME IS NOT NULL AND TABLE_NAME=? AND TABLE_SCHEMA=? AND CONSTRAINT_SCHEMA=? '
-            . 'GROUP BY CONSTRAINT_NAME '
             . 'ORDER BY TABLE_NAME ASC';
 
         $statement = $this->connection->prepare($query);
         $statement->setFetchMode(PDO::FETCH_ASSOC);
         $statement->execute([$tableName, $this->schema, $this->schema]);
 
+        $fksData = [];
+
+        // Prepare an array that groups foreign key data by constraint name
+        while ($row = $statement->fetch()) {
+            $constraintName = $row['CONSTRAINT_NAME'];
+
+            if (!isset($fksData[$constraintName])) {
+                $fksData[$constraintName] = [
+                    'constraint_name' => $constraintName,
+                    'local_table_name' => $row['TABLE_NAME'],
+                    'local_columns' => [],
+                    'foreign_table_name' => $row['REFERENCED_TABLE_NAME'],
+                    'foreign_columns' => [],
+                ];
+            }
+
+            $fksData[$constraintName]['local_columns'][] = $row['COLUMN_NAME'];
+            $fksData[$constraintName]['foreign_columns'][] = $row['REFERENCED_COLUMN_NAME'];
+        }
+
         $foreignKeys = [];
 
-        while ($row = $statement->fetch()) {
+        // Create the foreign keys
+        foreach ($fksData as $fkData) {
             $foreignKeys[] = new ForeignKey(
-                $row['CONSTRAINT_NAME'],
-                $row['TABLE_NAME'],
-                explode(',', $row['COLUMN_NAMES']),
-                $row['REFERENCED_TABLE_NAME'],
-                explode(',', $row['REFERENCED_COLUMN_NAMES'])
+                $fkData['constraint_name'],
+                $fkData['local_table_name'],
+                $fkData['local_columns'],
+                $fkData['foreign_table_name'],
+                $fkData['foreign_columns']
             );
         }